Add wrapped formatter funcs
These functions make it easier for the rust code to use the formatter hooks, since there is no more dealing with a `*mut *mut c_char` and all the pointers are already dereferenced, which means that the hook doesn't need any unsafe blocks (usually).

This commit also fixes some minor bugs that were undetected until now (i.e. Hook::FuncPrintDecorator had the wrong type and transmuted it which could lead to crashes).

The only thing left to do would be to decide whether we actually pass mutable reference to instances of ZydisDecodedInstruction and ZydisDecodedOperand. That would allow setting the user context, for which I don't really see a use case.